backend/requirements.txt
firebase_admin==7.1.0
Flask==3.1.2
python-dotenv==1.2.1
cloudinary==1.36.0
